requirements.txt
Django==2.2.9
pytz==2019.2
sqlparse==0.3.0
djangorestframework==3.11.0
